Integrate Prism to be able to parse Ruby Code (#31)
This pull request prepares `erbx` to be able to integrate `prism` for
parsing Ruby source code.

Since there isn't really a good way for compiling against prism from the
published gem on Rubygems and also not for compiling against the prism
gem that Ruby 3.4 ships/bundles with we are adding the `prism` gem as a
dependency to our Gemfile.

Since we specify the `prism` gem as a `github` source bundler will
actually clone the repo for us. In combination with the `.bundle/config`
file that configures bundler to install the gems into `vendor/bundle` we
get all sources files as if they were in the repo. This allows us to
compile prism from source so we can compile the shared library which in
turn erbx can use to compile itself against it.

This whole workflow is not really ideal, but it's the best way I was
able to figure out. I also explored using Git Submodules or inlining the
prism source code itself, but these approaches feel even hackier and
more error prone. With this current approach we can just leverage
bundler and it's capabilities which works everywhere where bundler
works.

---------

Co-authored-by: Seth Horsley <github.dvxhi@sl.isethi.com>