Parse ERB Nodes next to Text Content Nodes (#43)
This pull request improves the parsing of ERB tags next to text content
nodes. Additionally, new test cases have been added to ensure the
correct parsing of interpolated content within HTML elements.

I also renamed the `parser_parse_literal` to `parser_build_node` so it
can be used to build a node with a type and the name as a `char*`. Maybe
this should live in `ast_node.c` or replace `ast_node_init`, but leaving
this for now, since it's parser-specific enough.