Parse ERB inside HTML Comments (#44)
This pull request improves the parsing of ERB tags inside HTML comment
nodes. Additionally, a new test case has been added to ensure the
correct parsing of interpolated content within HTML comments.

We also changed the structure of the existing HTML comment nodes. The
content of the HTML comment itself is now stored in the `childrens`
array as literals or ERB Nodes.