Serialize Tokens as JSON and add `lex_json` command to CLI (#51)
This pull request adds the functionality to lex `.html.erb` files and to
serialize a `TOKEN_STRUCT` as a JSON string.

It also includes the new `lex_json` command in the CLI, and the
implementation of the `ERBX.lex_to_json` method in the Ruby bindings.

```
❯ ./erbx lex_json examples/simple_erb.html.erb
[{"type": "TOKEN_ERB_START", "value": "<%", "range": [0 , 2], "start": {"line": 1, "column": 0}, "end": {"line": 1, "column": 0}}, {"type": "TOKEN_ERB_CONTENT", "value": " title ", "range": [2 , 9], "start": {"line": 1, "column": 2}, "end": {"line": 1, "column": 2}}, {"type": "TOKEN_ERB_END", "value": "%>", "range": [9 , 11], "start": {"line": 1, "column": 9}, "end": {"line": 1, "column": 9}}, {"type": "TOKEN_NEWLINE", "value": "\n", "range": [11, 12], "start": {"line": 1, "column": 0}, "end": {"line": 1, "column": 0}}, {"type": "TOKEN_EOF", "value": "", "range": [12, 12], "start": {"line": 2, "column": 1}, "end": {"line": 2, "column": 1}}]
```

Additionally, it also implements `to_json` and `as_json` on both the
`ERBX::LexResult` and `ERBX::LibERBX::Token` Ruby classes.

In order to be able to implement this functionality, we also added the
`json_add_raw_string` and `json_add_size_t` functions in `json.c`.