Update fork helper to use `Tempfile` over `IO.pipe`
IO.pipe has a 65536-byte limit on macOS so if the Marshalled Ruby Object exceeds that limit the test gets stuck. Since we use minitest-difftastic it's quite likely that the Minitest::Result object with the whole difftastic diff can exceed that limit easily if you have a failing test that has a lot of changes between the two files we are comparing.