Expose `lex_file`, `parse_file`, `extract_html` and `extract_ruby` in Ruby Extension (#59)
This pull exposes the `lex_file`, `parse_file`, `extract_html` and
`extract_ruby` in the Ruby Extension.

We were now also able to enable the `extract_html` and `extract_ruby`
tests again.

Additionally, we refactored the way we build up the `LexResult` and
`ParseResult` Ruby objects, so we can share the implementation for the
regular `parse`/`lex` functions and the `parse_file`/`lex_file`
functions.