Add specialized error structs and error classes (#62)
This pull request adds support for specialized error structs in C and
error classes in Ruby - replacing the previous
`AST_UNEXPECTED_TOKEN_NODE_T` AST node that we used for any kind of
"error".

The `AST_NODE_T` base struct, which is used for all AST nodes, has a
`array_T* errors` field which now changed its "kind" from `AST_NODE_T`
to `ERROR_T`.

This allows us to simplify the construction of the error nodes in the
parser itself, since the parser now only has to pass over the right
arguments to the respective `*_error_init` function. Based on the type,
these functions know how to construct its error message and the
initialization of any special fields the error has.

It also allows to add the special metadata to each of the errors, which
is helpful, since sone errors need more or different context to what the
`AST_UNEXPECTED_TOKEN_NODE_T` struct could hold previously.

The definitions of the errors also live in the top-level `config.yml`
file.