Any element within `<svg>` can be void or non-void (#63)
This pull request enhances the parsing of HTML elements inside an
`<svg>` element. Within an SVG context, any tag name can be a regular
element (with opening and closing tag) or a self-closing element. But
all elements have to be closed.

**Valid:**
```svg
<svg>
  <path></path>
</svg>
```

**Valid:**
```svg
<svg>
  <path />
</svg>
```

**Invalid:**
```html
<svg>
  <path>
</svg>
```

This technically even applies to tag names that are void elements in the
HTML context:

**Valid (in SVG):**
```svg
<svg>
  <br></br>
</svg>
```

**Invalid (in HTML):**
```html
<br></br>
```