Simplify Prism integration by only building static library (#66)
This pull request simplifies the Prism integration and setup to only
build the static library. So we don't have to run `bundle install`
inside the Prism gem, don't have to compile the shared library and don't
have to compile the Prism Ruby extension.

With this change we also reduce the CI time for the build workflow from
~2m 45s down to ~1m.