JavaScript Bindings for the Browser and Node.js (#71)
This pull request introduces bindings for Javascript, both for the
Browser and Node.js.

### NX for managing JavaScript projects

This pull request introduces [`nx`](https://nx.dev) to help manage
workspaces so we can host multiple JavaScript projects in this repo. We
need multiple projects to build multiple independent NPM packages.

### NPM Packages

The pull request introduces the `javascript/packages` folder, which
holds the NPM packages which are going to be published on the
`@herb-tools` namespace on NPM.

For now we have these packages:
* **`@herb-tools/core`** - Core package which holds most of the classes
used that map to the structs in libherb.
* **`@herb-tools/browser`** - Specific Browser bundling and loading of
`libherb` for the browser (we still need a WASM build).
* **`@herb-tools/node`**- Specific Node.js bundling and loading of
`libherb` using a Node.js C++ Native Extension.

### Migration of the playground and docs apps

Since we now use `nx` we can also migrate the playground project to use
it and add a `package.json` for the `docs/` project so we can also
manage it using nx.


### Playground

The idea for the playground is to use the `@herb-tools/browser` package
so we can drop the express server and just parse everything in the
browser. For now, we use `@herb-tools/node` directly in the express app
instead of shelling out to the Ruby executable.

There are still some formatting issues, and AST conversion issues that
we can address later. We now also have new tabs for "Extracted Ruby",
"Extracted HTML", and the "Lex" output.

![CleanShot 2025-03-09 at 19 26
52@2x](https://github.com/user-attachments/assets/70286363-e854-42ca-a931-8d9a83733db6)

### libherb

The Makefile in the main project now also builds a static library so the
Node.js C++ native extension can build against it.