Add WASM build for libherb using emscripten (#72)
This pull request adds a build of `libherb` in the `wasm/` directory
that uses emscripten (via `emcc` and `em++`) to build a WebAssembly
(WASM) module.

This allows us to utilize the build in the `@herb-tools/browser` package
to run the parser fully in the browser.

Additionally, this pull request also updates the playground to use the
`@herb-tools/browser` package which now eliminates the need for a
server/API to do the parsing in the playground.

![CleanShot 2025-03-12 at 20 34
48@2x](https://github.com/user-attachments/assets/d2759c3e-0c67-446a-af91-728ac8d45842)