Improve JavaScript `HerbBackend` interface by introducing `load()` (#75)
This pull request improves the `HerbBackend` interface in the JavaScript
packages.

The `@herb-tools/core` now provides an abstract `HerbBackend` class that
both the `@herb-tools/browser` and `@herb-tools/node` packages
implement.

The key difference is that all functions are now not async-anymore,
making it nicer to interact with them.

Before, all functions were async and were always loading the backend for
each function call. This was actually noticable because it would re-load
the whole WASM backend on every keystroke.

Now with introducing one single asynchronous `load()` function, that the
caller has to call once, we can load the `LibHerbBackend` and store the
resolved backend on the `HerbBackend` instance, making subsequent calls
more efficient.