Fix and improve location information for newline tokens (#76)
This pull request improves the way the lexer handles newlines, so we get
correct location and position information for newline tokens.

Previously, the `start` location of the `TOKEN_NEWLINE` was always set
to `x:0`, which isn't always correct. This pull request now moves the
logic for advancing the current_line to the `token_init` function and
tracks it using the new `previous_*` fields on the lexer struct.

Technically this logic isn't quite right in the `token_init`, but since
we don't create new tokens from anywhere else we are going to leave it
there for now.