Use zero-based columns in token positions (#77)
This pull request changes the columns in token positions to be
zero-based throughout. The first line was already using a zero-based
column, but any following line restarted at column 1, which made the
column numbers inconsistent.

So, this pull request updates the columns to be zero-based for all
lines, following the way [Prism](https://github.com/ruby/prism) is
handling it. This is intentional, even when quite a few text editors are
using 1-based column numbers.

Line numbers stay 1-based.