Use Monaco Editor in playground to render diagnostics and to sync editor with pretty output (#78)
This pull request includes significant updates to the playground by
introducing the Monaco editor.

Using the Monaco editor, we can show the parse errors in the editor.

Additionally, this pull request implements a two-way sync between the
editor on the left and the pretty output on the right. By clicking
anyway in the editor it will select and jump to the corresponding node
in the tree.

Clicking or hovering over any Node, Error, Token or location in the
pretty output highlights the corresponding text in the editor on the
left.


https://github.com/user-attachments/assets/4f663f0b-4d46-44c9-bd26-346ee62523a6