Parse, analyze and transform ERB nodes to common ERB control flow structures (#80)
This pull request adds an analyze/transform step so that a
`ERBContentNode` can be represented more accurately and meaningfully
with specific Ruby control flow nodes.

So for example, where we have:

```html+erb
<% if true %>
  Content
<% end %>
```

It previously would put the ERB tags on the same level as the text
content node, which is not quite right, semantically:

```javascript
@ DocumentNode (location: (1:0)-(3:9))
├── errors: []
└── children: (3 items)
    ├── @ ERBContentNode (location: (1:0)-(1:13))
    │   ├── errors: []
    │   ├── tag_opening: "<%" (location: (1:0)-(1:2))
    │   ├── content: " if true " (location: (1:2)-(1:11))
    │   └── tag_closing: "%>" (location: (1:11)-(1:13))
    │   
    ├── @ HTMLTextNode (location: (1:13)-(3:0))
    │   ├── errors: []
    │   └── content: "\n  Content\n"
    │   
    └── @ ERBContentNode (location: (3:0)-(3:9))
        ├── errors: []
        ├── tag_opening: "<%" (location: (3:0)-(3:2))
        ├── content: " end " (location: (3:2)-(3:7))
        └── tag_closing: "%>" (location: (3:7)-(3:9))
```

So this pull request changes this by parsing the ERB content as Ruby
code. If a single ERB content node contains valid syntax, we leave the
`ERBContentNode` untouched, since it's self contained. But if we have a
split up Ruby statement we try to collect up the nodes and put them in
the corresponding control structure.

So the above example now gets parsed as:

```javascript
@ DocumentNode (location: (1:0)-(3:9))
├── errors: []
└── children: (1 item)
    └── @ ERBIfNode (location: (1:0)-(3:9))
        ├── errors: []
        ├── tag_opening: "<%" (location: (1:0)-(1:2))
        ├── content: " if true " (location: (1:2)-(1:11))
        ├── tag_closing: "%>" (location: (1:11)-(1:13))
        ├── statements: (1 item)
        │   └── @ HTMLTextNode (location: (1:13)-(3:0))
        │       ├── errors: []
        │       └── content: "\n  Content\n"
        │       
        │   
        ├── subsequent: ∅
        └── end_node: 
            └── @ ERBEndNode (location: (3:0)-(3:9))
                ├── errors: []
                ├── tag_opening: "<%" (location: (3:0)-(3:2))
                ├── content: " end " (location: (3:2)-(3:7))
                └── tag_closing: "%>" (location: (3:7)-(3:9))
```

This means we introduced new AST ERB nodes for `if/elsif/else/end`,
`case/when/else/end`, `while/end`, `for/end`, `until/end`,
`unless/else/end`, `begin/rescue/else/ensure/end`, `do/end`, or `{ }`
blocks.

The design of these nodes is mirroring the design of the equivalent
Prism AST node. Maybe we can actually use the Prism AST nodes in the
future or make them 1:1 compatible. For now, they are close enough.

Something I'd love to add in the future is to mirror the exact fields
from the Prism AST, so that you know, for example, for the `when` node,
what the parsed `condition` are. We would have this information right
now, but we don't have an elegant way to serialize/map the foreign Prism
Nodes in the current setup.

Additionally, we extract the whole Ruby code out of the HTML+ERB source,
and parse it with Prism. Any Ruby parse or syntax errors get appended as
`RubyParseError` to the HTML+ERB ParseResult.

But in order to be able to parse the Ruby code in any HTML+ERB source we
need to append semicolons to each ERB Content Node. Because a template
like this:

```html+erb
<% if true %>A<% else %>B<% end %>
```
would be extracted as:
```ruby
   if true       else       end   
```

which is not valid Ruby, since if statements have to be broken up to
multiple lines. So we have to append semicolons to correctly terminate
the individual statements:

```ruby
   if true ;     else ;     end ; 
```

And with this, we can now also parse single-line Ruby statements,
preserve the exact location information of the Ruby code as it was in
the HTML+ERB source, which means we can use the Prism errors 1:1 without
having to remap the location information.