Add RBS Inline Type Annotations (#106)
This pull request adds RBS Inline Type annotations for the Ruby gem in a
first iteration. These are not perfect yet, but are a good starting
point. We are also running `steep check` in CI now.

I've also been trying to get Sorbet with RBS comments to work but sadly
it doesn't support RBS aliases yet.

```
$ bundle exec srb tc

lib/herb/errors.rb:21: RBS aliases are not supported https://srb.help/3551
    21 |      #: () -> serialized_error
                       ^^^^^^^^^^^^^^^^
```

The only other option is to inline the whole Hash type which didn't seem
like a great option, because the would a) be super hard to render out
and b) it's super verbose. Which is why we don't run `srb tc` in CI yet.
But I'd love to get that to get that working too in the future.

Resolves #99