Rework result classes and types in JavaScript packages (#125)
This pull request reworks the result classes and types in the
`@herb-tools/core` JavaScript package.

The `Result` class is now abstract, the `success` method was renamed to
`successful` and both the `successful` and `failed` methods are now
getters.

We now also export `HerbError` and `HerbWarning` from the
`@herb-tools/core` package, so they can be used in other packages.

Finally, we removed some of the leftover `any` types and migrated them
to use proper types.