Herb Language Server and Visual Studio Code Extension (#105)
This pull request adds a Herb Language Server Protocol implementation
and a Visual Studio Code Extension for it.

To be able to make the bundling work in CommonJS in VS Code I also
implemented a new `@herb-tools/node-wasm` package that uses the
WebAssembly build in `wasm/` to build a version of the libherb library
for the Node.js environment.

Ideally we can use the Node.js implementation in the future, but the
pre-compilation of the native Node.js Addon is trickier than initially
expected with Electron and the system architecture.