Add `Diagnostics` interface and rename other to `MonacoDiagnostic` (#229)
This pull request introduces a new `Diagnostics` interface which will be
used to represent errors, warnings, info messages, lint offenses, and
anything else that could be shown in the editor as a "Diagnostic".

Previously we had a `HerbDiagnostic` type, which we used to convert
"errors" to "diagnostics" so they would be compatible with the Monaco
editor in the playground.

Now we renamed that `HerbDiagnostic` type to `MonacoDiagnostic` so it's
clear what this one is used for. With that, we also introduced a new
`toMonacoDiagnostic(diagnostic: Diagnostic): MonacoDiagnostic` utility
function to convert any `Diagnostic` to a `MonacoDiagnostic`.

There are probably also some refactoring opportunities to make use of
these new types in the Language Server.