Show linter offenses inline in rule documentation pages (#233)
This pull request adds the functionality to see the linter offenses
right in the code blocks in the documentation. This is especially useful
on the rule documentation pages since it immediately shows what's
"wrong" with the given snippet.

When building the documentation site we look for `erb` code blocks and
transform them by running the Herb Linter on the code snippet. We then
pass the code and the offenses to [Shiki](https://shiki.style) and
[Twoslash](https://twoslash.netlify.app) to render the code snippet with
syntax highlighting and rendering the diagnostics inline.

![CleanShot 2025-07-04 at 12 11
45@2x](https://github.com/user-attachments/assets/33dfa2b6-fcca-463c-a7a3-05f193131c5f)

![CleanShot 2025-07-04 at 12 11
33@2x](https://github.com/user-attachments/assets/ef4c3112-8f14-4a64-af93-482fca441da5)

Additionally, it refactors the `config.mts` by extracting most of the
logic into helper/config files and updates some of the Linter messages.