Herb Highlighter (#234)
This pull request introduces a new `@herb-tools/highlighter` package,
which can be used to syntax highlight HTML+ERB in the terminal, render
code snippets, and render diagnostics.

It was primarily built to simplify the Herb Linter CLI. That CLI has to
render diagnostics by rendering sections of a file and annotate that
snippet with the lint offenses.

The `herb-highlight` CLI can also be used standalone (almost like a
`cat` alternative) for HTML+ERB files. This is mostly for debugging
purposes.

![CleanShot 2025-07-04 at 15 00
35@2x](https://github.com/user-attachments/assets/ac36a871-00f4-437b-a25f-e8c89d051fa6)

The Highlighter comes with a few default themes and has the option to
use custom themes too (using a Theme JSON file, maybe we can in the
future rely on a more well-known format).

![CleanShot 2025-07-04 at 15 02
28@2x](https://github.com/user-attachments/assets/c602a864-fc02-4f8a-ba7c-25519bfdcfe9)

It also has option to focus on lines (and render the surrounding lines
with dimmed colors):

![CleanShot 2025-07-04 at 15 03
08@2x](https://github.com/user-attachments/assets/d61391d1-ae38-435f-bb2e-5b65a81b186f)

You can also show more lines around the focused line using
`--context-lines`:

![CleanShot 2025-07-04 at 15 04
15@2x](https://github.com/user-attachments/assets/45b00ddc-9ac5-4945-81eb-0efc67d87611)


There are a lot more options:
```
❯ bin/herb-highlight --help
Usage: herb-highlight [file] [options]

Arguments:
  file             File to highlight (required)

Options:
  -h, --help       show help
  -v, --version    show version
  --theme          color theme (onedark|github-light|dracula|tokyo-night|simple) or path to custom theme file [default: onedark]
  --focus          line number to focus on (shows only that line with context)
  --context-lines  number of context lines around focus line [default: 2]
  --no-line-numbers hide line numbers and file path header
  --wrap-lines     enable line wrapping [default: true]
  --no-wrap-lines  disable line wrapping
  --truncate-lines enable line truncation (mutually exclusive with --wrap-lines)
  --max-width      maximum width for line wrapping/truncation [default: terminal width]
```

For the future it would be cool to also support rendering to HTML so we
could use it in the browser.

For the Herb Linter this now means that it gets syntax highlighting:

![CleanShot 2025-07-04 at 15 07
10@2x](https://github.com/user-attachments/assets/24e8188e-9b1a-4fac-b185-a2034f803734)

Some of the highlight options are always available in the Herb Linter
CLI, like `--theme`, `--no-wrap-lines`, `--truncate-lines`, etc.