Herb Formatter (#192)
This pull request introduces a new package, `@herb-tools/formatter`,
which provides utilities for formatting HTML+ERB templates.

It includes the implementation of the formatter logic, adds a CLI to
test the formatting using `bin/herb-formatter [file]`, implements the
Formatter into the language-server for format-on-save support, and adds
a new tab to the playground to play with the formatter in the
browser-based playground.

Note: the current state of this is not 100% ready yet, as it formats a lot of
documents in a sub-optimal way.

![CleanShot 2025-06-29 at 18 23
56@2x](https://github.com/user-attachments/assets/2c04f652-605c-46fd-91c2-df0e102a1b4d)


https://github.com/user-attachments/assets/e1c22310-38dd-4fd8-8702-3a1ca06c7659

![CleanShot 2025-06-29 at 18 19
52@2x](https://github.com/user-attachments/assets/651f6a4f-719d-4f37-a15a-95b59e263ba2)

![CleanShot 2025-06-29 at 18 20
52@2x](https://github.com/user-attachments/assets/65999cba-ff72-4e41-bbf2-f1bfcedaf195)

---------

Signed-off-by: Marco Roth <marco.roth@intergga.ch>