Allow HTML Attribute Names to have a period in the attribute name (#324)
This pull request adds support for parsing a period in HTML attribute
names. So we can now properly parse snippets like this:

```erb
<div x-transition.duration.500ms></div>
```

Technically this is not fully compliant HTML5, but popular frameworks
like Alpine.js leverage this syntax and it mostly seem to work in
browsers.

But even if we wouldn't want to allow this kind of syntax, we'd still
need to parse it this way so we could write linter rules to help people
avoid this kind of syntax if strict HTML5-compliance is required.

Resolves https://github.com/marcoroth/herb/issues/287