Linter: Allow a new `LintContext` to be passed to linter rules (#340)
This pull request adds a context to the linter rule to be able to
distinguish between file linting and snippet linting. Rules can now
access filename information through a shared context object, enabling
different behavior for actual files versus code snippets used in
documentation or testing.

We added this functionality so we could remove the
`erb-requires-trailing-newline` offenses in all the code snippets in the
docs, that weren't really valid since the snippets in markdown aren't
really files. By not passing a filename we now assume that this comes
from stdin or a snippet and therefore don't the
`erb-requires-trailing-newline` rule in that case.

This change also allows us to pass around context in the future to build
more sophisticated rules.