Language Server: Add `excludedRules` configuration (#363)
This pull request adds configurable `excludedRules` support to the
language server and VSCode extension.

Since #362 introduced the `parser-no-errors` rule we now have two ways
of surfacing the parser errors in the language server.

The language server now allows to configure which linter rules to
exclude from diagnostics. The `parser-no-errors` rule is set as a
default to prevent duplicate parsers errors in the editor.

But this change now also allows users to exclude other linter rules they
might want to disable (just in the language server for now):


https://github.com/user-attachments/assets/1cf108b4-856e-4a96-b372-4a4e1f678a6d