Parser: Don't crash on long ERB comments (#376)
This pull request updates the parser from analyzing the ERB comments as
Ruby code. This was causing stack overflow errors in WASM environments
and Unicode character corruption in the formatter.

Technically we didn't really solve the underlying issue, which seems to
be that Prism is crashing in the WASM build when given such a "deeply
nested" structure to parse.

Given this input: 
```erb
<%# x x x x x x x x x x x %>
```

means that we are passing this Ruby Code to Prism:
```ruby
x x x x x x x x x x x 
```

Which seems to make it crash in the WASM environment. I opened #377 to
investigate/resolve that.

So for now, we are just not analyzing ERB comments, since we don't get
any value from that anyway which also solves the most common way how
this was surfaced.

Resolves https://github.com/marcoroth/herb/issues/247
Resolves https://github.com/marcoroth/herb/issues/292

Might also resolve https://github.com/marcoroth/herb/issues/283 and
https://github.com/marcoroth/herb/issues/350