Convert field `name` on `HTMLAttributeNameNode` to a `children` array (#189)
This pull request changes the parsing logic for HTML attribute names and
updates the `name` field of the `HTMLAttributeNameNode` to a `children`
field with type `array`.

This allows us to store an array of nodes, like we do for the value of
the `HTMLAttributeValueNode`. With this, we can support more complex
attribute structures, i.e interpolations in the HTML attribute name,
like:

```html+erb
<img <%= key %>="true">
<img data-<%= key %>-name="example">
<img data-<%= key %>="example">
<img <%= key %>-value="example">
```

Even if this is quite discouraged syntax, it allows us to parse it so we
could write a linter rule to steer people away from using this syntax 🙈.

This also updates the linter and the affected linter rules, by updating
the `AttributeVisitorMixin` to support all combinations of
static/dynamic key/values and also adding new AST Utils in the
`@herb-tools/core` package.

Resolves #85
Resolves #387