Implement `@herb-tools/printer` package for printing AST nodes (#416)
This pull request introduces a new `@herb-tools/printer` package for
printing Herb Syntax Tree Nodes without any formatting.

The use-case for this is to be able to 
* exactly print back any Node to source code (i.e for showing them in
Linter error messages)
* parse a file, modify the nodes/content, and print it back to source
without loosing any details and not formatting anything else in the
file.
* implement auto-fix actions for Linter Rule offenses
* and probably more.

The `Printer` expects the source to be parsed using the
`track_whitespace: true` parser option.