Formatter: Refactor Formatter to use `@herb-tools/printer` (#423)
This pull request refactors the `Printer` in the formatter as
`FormatPrinter` and refactors the implementation to use the new
infrastructure provided by the `@herb-tools/printer` package introduced
in #416.

Additionally it marks some of the `FormatPrinter` methods and fields as
`@depracted` so we can start using the infrastructure provided in the
`@herb-tools/printer` package.