Linter: Add `isEnabled` hook for `LinterRule` (#430)
This pull request implements a new optional `isEnabled` API for any
`LinterRule` to implement. The linter rule itself can decide if it
should be disabled given the `ParseResult` and the `LintContext`.

```ts
class MyRule extends ParserRule {
  name = "my-rule"

  isEnabled(result: ParseResult, context?: Partial<LintContext>): boolean {
    // Only run on .erb files
    if (!context?.fileName?.endsWith('.erb')) return false

    // Only run if document has specific nodes
    return result.value.children.some(child => isNode(child, HTMLElementNode))
  }

  check(result: ParseResult, context?: Partial<LintContext>): LintOffense[] {
    // Rule logic here
  }
}
```

This pull request also updates the `html-tag-name-lowercase` to make use
of this new API. It will disable itself if either the filename ends in
`.xml` or `.xml.erb` of if the document itself contains a
`XMLDeclarationNode`.