Printer: Move `IdentityPrinter` implementation out of `Printer` (#435)
This pull request moves the concrete implantation of the
`IdentityPrinter` into the `IdentityPrinter` class, so that the abstract
`Printer` super class only handles the visiting/printing.

The idea is that you should overwrite the methods for the nodes you want
to handle in your `Printer`. If you don't overwrite anything it
shouldn't print anything.

If you want to just fallback to the default, you can extend from the
`IdentityPrinter` class and then overwrite the methods you want to
handle differently.