Formatter: Add support for Content Preserving Elements (#438)
This pull request improves the way the formatter handles Content
Preserving Elements like `<pre>`, `<textarea>`, `<style>` and
`<script>`.

These elements have different-requirements when it comes to whitespace,
so for now, we just keep them as-is so we don't mis- or over-format them
and break some of the application logic.

This also partially addresses #436.