Printer: Also fully reconstruct whitespace in `HTMLCloseTag` (#443)
This pull request updates the `IdentityPrinter` to also reconstruct the
whitespace in the `HTMLCloseTag`.

The parser was already adding the `WhitespaceNode`s to the `children`
array of the `HTMLCloseTag` (done in #396), but the `IdentityPrinter`
wasn't accounting for them.

Now these templates can also be fully reconstructed using the
`IdentityPrinter`:
```html
<div></   div>
<div></div   >
<div></   div   >
```

Additionally, this pull request also adds some more AST helpers to the
`@herb-tools/core` package to select nodes before/after a certain
`Location` or `Position`.