Implement `stimulus-lint` package using Herb Parser and Linter (#457)
This pull request implements an early version of the `stimulus-lint`
package using the Herb Parser, Herb Linter, Stimulus Parser and by
reusing and building on top of the Herb Linter CLI, which means we
automatically get all of the features the Herb Linter provides.

In the future, we want to extract all of the diagnostics the Stimulus
LSP currently provides in the Stimulus LSP codebase itself and migrate
them as Linter Rules for `stimulus-lint`. That way, the rules can be
used independently of the Stimulus LSP. The Stimulus LSP can then
integrate `stimulus-lint` to show these diagnostics, similar to how the
Herb LSP and Herb Linter are setup.

This package might get extracted to
`http://github.com/marcoroth/stimulus-lint` at a later point, but for
now it's easier to iterate as we have the dependent packages right next
to it.

<img width="5056" height="2804" alt="CleanShot 2025-08-28 at 02 38
23@2x"
src="https://github.com/user-attachments/assets/63b1813b-38c4-460d-91f7-bfbcd6044563"
/>

<img width="4958" height="2570" alt="CleanShot 2025-08-28 at 02 40
23@2x"
src="https://github.com/user-attachments/assets/86138582-01a1-4ec1-b09a-6667bf3fae2d"
/>

(The reporting at the end of a run still needs some work)