Parser: Don't treat guard clauses as `ERBIfNode` or `ERBUnlessNode` (#517)
This pull request updates the parser and the analyzer to only transform
an `ERBContentNode` node to an `ERBIfNode` or `ERBUnlessNode` when it
has both `if/unless` and `end` keywords.

Guard clauses or ternaries shouldn't be treated as `ERBIfNode` or
`ERBUnlessNode`. We might want to revisit this when we expose the actual
prism nodes for each `ERBContentNode`, but for the sake of
parsing/formatting we shouldn't treat those `ERBIfNode` or
`ERBUnlessNode` nodes for now.

Resolves #505