Parser: Namespace Parser C functions (#534)
As painstakingly worked out with @marcoroth during Kaigi on Rails, macOS
26 Tahoe introduces a function named `parser_init` that can clobber the
Herb function with the same name. Precompiled gems no longer crash on
macOS 26 after this change to ensure the Herb functions will not be
clobbered by the OS functions.

From what we could tell, macOS has a new(?) framework called "Biome" which both defines `parser_parse` and `parse_init` functions. 

We believe the reason why this bug only occured in the precompiled version of the gem is because at compile time it wouldn't know about both of these functions. So at runtime  it would "choose" the first matching one, which happened to be the wrong one, thus messing with the stack/structs when calling the function here:

https://github.com/marcoroth/herb/blob/b93eba9f80e7828428e084efd103acdf77bee81a/src/herb.c#L32

which then lead to this `EXC_BAD_ACCESS (code=1, address=0x29)` error later in: 

https://github.com/marcoroth/herb/blob/b93eba9f80e7828428e084efd103acdf77bee81a/src/parser.c#L1210

Resolves https://github.com/marcoroth/herb/issues/484
Resolves https://github.com/marcoroth/reactionview/issues/13

Co-Authored-By: Marco Roth <marco.roth@intergga.ch>